requirements/testing.txt
-r production.txt
pytest==6.2.5
pytest-cov==2.12.1
pytest-aiohttp==0.3.0
pytest-asyncio==0.18.3
flake8==4.0.1
flake8-isort==4.1.1
pydocstyle==6.1.1
pylint==2.13.7
pylint-strict-informational==0.1
setuptools==62.1.0
tox==3.25.0
mypy==0.942
pre-commit==2.18.1